Extract ASCII validation into a separate file
