+++ src/lasx/implementation.cpp
@@ -148,6 +148,7 @@ convert_utf8_1_to_2_byte_to_utf16(__m128i in, size_t shufutf8_idx) {
 #include "generic/buf_block_reader.h"
 #include "generic/utf8_validation/utf8_lookup4_algorithm.h"
 #include "generic/utf8_validation/utf8_validator.h"
+#include "generic/ascii_validation.h"
 
 // transcoding from UTF-8 to Latin 1
 #include "generic/utf8_to_latin1/utf8_to_latin1.h"
@@ -212,12 +213,12 @@ simdutf_warn_unused result implementation::validate_utf8_with_errors(
 
 simdutf_warn_unused bool
 implementation::validate_ascii(const char *buf, size_t len) const noexcept {
-  return lasx::utf8_validation::generic_validate_ascii(buf, len);
+  return lasx::ascii_validation::generic_validate_ascii(buf, len);
 }
 
 simdutf_warn_unused result implementation::validate_ascii_with_errors(
     const char *buf, size_t len) const noexcept {
-  return lasx::utf8_validation::generic_validate_ascii_with_errors(buf, len);
+  return lasx::ascii_validation::generic_validate_ascii_with_errors(buf, len);
 }
 
 simdutf_warn_unused bool
